// frontend/src/client/types.gen.ts
export type UserProfileResponse = {
    /**
     * Id
     */
    id: string;
    /**
     * Email
     */
    email: string;
    /**
     * Username
     */
    username: string;
    /**
     * Role
     */
    role: string;
    /**
     * Is Active
     */
    is_active: boolean;
    /**
     * Is Verified
     */
    is_verified: boolean;
    /**
     * Created At
     */
    created_at: string;
    /**
     * Last Login At
     */
    last_login_at?: string | null;
};